ecommerce_proj/E-commerce_App-main/Backend/flask_backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
psycopg2-binary==2.9.7
gunicorn==20.1.0
marshmallow==3.19.0
python-dotenv==1.0.0
bcrypt==4.0.1
Werkzeug==2.3.7
elasticsearch==8.10.0
alembic==1.12.0
